microservices/checkout/requirements.txt
Flask>=1.0.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-cloud-logging==1.14.0
google-cloud-error-reporting==0.33.0
gunicorn==19.9.0
six==1.11.0
sendgrid
requests
werkzeug==0.16.1
flask_oauthlib
Flask-Login
Flask-Session
datetime
flask-cors
pyjwt
